requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-filter>=23.5
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
black>=23.11.0
PyYAML>=6.0  # For YAML fixture file support
# Performance optimization dependencies
django-debug-toolbar>=4.2.0  # For development performance monitoring
django-extensions>=3.2.0  # For additional Django management commands